Crate xfr

An atomic ring for processing memory-mapped frames.

This crate provides two primary data structures Ring and FrameBuffer. Buffers hold an array of frames and rings provide producer/consumer mechanics for implementing packet flows. The benefit of this combination of data structures is that to implement packet flows, data frames do not need to be moved or copied. Ring elements contain pointers to frames and ring elements are what’s moved around to implement packet flow. This general design is heavily inspired by Linux’s eXpress Data Path (XDP). Xfr Rings are lock-free providing a higher-performance more robust substrate than lock based rings.

Basic packet flow

Let’s start with a simple packet flow example. In this scenario we have two data link interfaces. For each data link interface there is an external i/o driver that is responsible for taking packets off the wire and placing them in a buffer for further processing. The i/o driver is also responsible for taking buffered packets that are ready for egress and putting them on the wire. Packets transit between these two interfaces through a packet processing engine.

The interface i/o drivers and the packet processing engine interact through a set of rings that provide access into an underlying frame buffer. In the diagram below each interface is connected to an ingress ring (upper) that it uses to buffer packets that are ready for further processing. Each interface is also connected to an egress ring (lower) that it consumes packets from that are ready to be put on the wire.

                       h
                       t     r
    *------* produce  *=======================* consume  *--------*
    |      |--------->|0A|0B|0C|  |  |  |  |  |--------->|        |
    | ifx0 |          *=======================*          |        |
    |      |  consume *=======================*  produce |        |
    |      |<---------|  |  |  |  |  |  |  |  |<---------|        |
    *------*          *=======================*          |        |
                                                         |        |
                      *=======================*          |        |
                      |00|01|02|03|04|05|06|07|          |        |
                      *-----------------------*          |        |
                      |08|09|0A|0B|0C|0D|0E|0F|          | Packet |
                      *-----------------------*          | Engine |
                      |10|11|12|13|14|15|16|17|          |        |
                      *-----------------------*          |        |
                      |18|19|1A|1B|1C|1D|1E|1F|          |        |
                      *=======================*          |        |
                                                         |        |
    *------* produce  *=======================* consume  |        |
    |      |--------->|  |  |  |  |  |  |  |  |--------->|        |
    | ifx1 |          *=======================*          |        |
    |      |  consume *=======================*  produce |        |
    |      |<---------|  |  |  |  |  |  |  |  |<---------|        |
    *------*          *=======================*          *--------*

The ingress and egress rings do not contain any packet frame data. Rather they contain an index into a frame buffer that holds packets while they are being processed. The act of packet processing happens over these frame-pointer ring elements. This means we do not need to move packet data around as it is processed.

In the example above the producer has 3 frames to produce for processing. The first step it must take is reserving 3 frame pointers from the ingress ring buffer. This moves the rsvd index forward 3 positions. The interface driver can now modify frame pointers between the head index and the rsvd index. This is depicted in the diagram above by the h and r markers.

Once the interface driver has a reserved region in the ingress ring buffer, it must allocate frames from the frame buffer (depicted in the center of the diagram). In this example the interface driver allocates 3 frames from the frame buffer and it gets back frame addresses 0A, 0B and 0C. It then assigns these addresses to the frame pointer ring elements it has reserved. This is also depicted in the diagram above.

The interface driver can now write packets to the frame buffer through the ring buffer pointers it owns. Once it has done that, it calls produce(3) on the ingress ring. This results in the head index moving forward 3 positions.

                            h
                      t     r
   *------* produce  *=======================* consume  *--------*
   |      |--------->|0A|0B|0C|  |  |  |  |  |--------->|        |
   | ifx0 |          *=======================*          | Packet |
   |      |  consume *=======================*  produce | Engine |
   |      |<---------|  |  |  |  |  |  |  |  |<---------|        |
   *------*          *=======================*          |  ...   |

This results in a space of 3 frame pointers between the tail depicted as the marker t in the diagrams, and the head h. The space between these two indices is the consumable space. This means that the packet engine now has ownership of these frames e.g. the act of producing frames is transfer of ownership from the producer to the consumer. When the packet engine has finished processing the frames, it can call consume(3) to move the tail index forward 3 frame pointers. At this point, assuming there have been no new packets produced, there are no frames to produce and no frames to consume on the ingress ring.

                            t
                            h
                            r
   *------* produce  *=======================* consume  *--------*
   |      |--------->|0A|0B|0C|  |  |  |  |  |--------->|        |
   | ifx0 |          *=======================*          | Packet |
   |      |  consume *=======================*  produce | Engine |
   |      |<---------|  |  |  |  |  |  |  |  |<---------|        |
   *------*          *=======================*          |  ...   |

Now how about egress? The process works the same in reverse on the egress ring. Here the packet engine is the producer and the interface driver is the consumer.

                            t
                            h
                            r
   *------* produce  *=======================* consume  *--------*
   |      |--------->|0A|0B|0C|  |  |  |  |  |--------->|        |
   | ifx0 |          *=======================*          | Packet |
   |      |  consume *=======================*  produce | Engine |
   |      |<---------|  |  |  |  |  |0C|0B|0A|<---------|        |
   *------*          *=======================*          |  ...   |
                                     r      h
                                            t

Here the packet engine starts by reserving 3 frame pointers on the egress ring as depicted by the r marker in the diagram above. Once reserved it writes the indices of the frames it has processed in those three frame pointer elements within the ring. Then it moves the head pointer forward with a produce(3) call

                            t
                            h
                            r
   *------* produce  *=======================* consume  *--------*
   |      |--------->|0A|0B|0C|  |  |  |  |  |--------->|        |
   | ifx0 |          *=======================*          | Packet |
   |      |  consume *=======================*  produce | Engine |
   |      |<---------|  |  |  |  |  |0C|0B|0A|<---------|        |
   *------*          *=======================*          |  ...   |
                                     r      t
                                     h

The interface driver now owns the frames between the head and tail indices. It takes the referenced frames and puts them on the wire. Then it calls consume(3) resulting in the tail index catching up with head.

                            t
                            h
                            r
   *------* produce  *=======================* consume  *--------*
   |      |--------->|0A|0B|0C|  |  |  |  |  |--------->|        |
   | ifx0 |          *=======================*          | Packet |
   |      |  consume *=======================*  produce | Engine |
   |      |<---------|  |  |  |  |  |0C|0B|0A|<---------|        |
   *------*          *=======================*          |  ...   |
                                     r
                                     h
                                     t

Tada! We’ve processed a pile of packets without copying any frame data. The rings are a thread-safe mechanism for producers and consumers of packet frames to exchange ownership of frames through pointers.

Ring invariants

The Ring data structure is able to work in a lock-free way because of a few key invariants, combined with a notion of epochs. When an index rolls over to the beginning of the ring it enters a new epoch of that ring. The ring invariants can be summarized in the following inequality.

t <= h <= r <= t+R

Taking each of thse in turn

t <= h: The tail can never eclipse the head.

h <= r: The head can never eclipse the reserve.

r <= t+R: The reserve can never eclipse the tail in the next epoch.

The last constraint also insures that the indices span no more than two epochs.

When we want to compare two indices we need to know their reltaive position within the ring e.g., the value modulo the size of the ring, as well as what epoch the indices are in. Combining the facts that a) there is a complete ordering across indices, b) and indices can only span a maximum of two epochs - this allows us to track epochs as a single bit quantity. All we need to know to compare two indicies is whether they are in the same or different epochs. Which specific indices are being compared allows us to determine which index is in the latter epoch. For example if we are comparing t and h and they are in different epochs, we know that h is in the later epoch because of the invariant t <= h.

Epoch Encoding

The single-bit representation of an epoch allows for a very efficient index encoding scheme that can be read and written to atomically. Being able to operate atomically on indices and their epochs is critically important for achieving a lock free ring.

Consider a ring consumer enforcing the invariant t <= h. Only the consumer can move t, but a producer in another thread can move h. So if the consumer reads h at time a and then reads the epoch for h at time b but h moved between a and b, then the total representation of h collected (its value and its epoch) is invalid. To get a consistent total representation of h we would need to lock the index and epoch for h.

However, because the epoch is just a one-bit quantity, we use the leading bit of an index to encode an epoch. Doing this allows us to atomically load the index and it’s epoch. Ring indices are 64 bit values, so this effectively makes them 63 bit values. I don’t think we’ll be in need of Gangam style rings exceeding 9 quintillion elements any time soon.

Usage

The following example demonstrates transferring 1.5 GB of data in 1500 byte frames across a memory mapped ring.

use std::sync::Arc;
use std::thread::{spawn, sleep};
use std::time::Duration;
use std::str::from_utf8;

use xfr::*;

use rand::Rng;
// Transfer 1,000,000 1500 byte frames (1.5 GB)
const R: usize = 1024;
const N: usize = 4096;
const F: usize = 1500;
const K: u32 = 1000000;

let fb = Arc::new(FrameBuffer::<N, F>::new());
let (p, c) = ring::<R, N, F>(fb);

let t1 = spawn(move|| {
// create test data we're going to send the full 1500 bytes per
// packet, but only bother writing a 4 byte integer in each packet.
let mut data = Vec::new();
for i in 0..K {
    data.push(i.to_be_bytes());
}

let mut rng = rand::thread_rng();
let mut i = 0;
loop {
    let count = usize::min(rng.gen_range(0..10), (K-i) as usize);
    let fps = match p.reserve(count) {
        Ok(fps) => fps,
        Err(_) => continue,
    };
    for fp in fps {
        p.write(fp, data[i as usize].as_slice());
        i += 1;
    }
    p.produce(count).unwrap();
    if i >= K {
        break;
    }
}
println!("producer finished");
});

let t2 = spawn(move|| {
let mut total = 0u32;
loop {
    let mut count = 0;
    let consumable = c.consumable();
    for fp in consumable {
        let content = c.read(fp);
        let x = u32::from_be_bytes(content.try_into().unwrap());
        assert_eq!(x, total);
        total += 1;
        count += 1;
    }
    if count == 0 {
        continue
    }
    c.consume(count).unwrap();
    if total >= K {
        break;
    }
}
println!("consumer finished");
});

t1.join().unwrap();
t2.join().unwrap();

Structs

Frame

A Frame is a wrapper for a network data frame. The data inside is owned by this frame.

FrameBuffer

A frame buffer holds an array of frames and an index that identifies the first allocatable frames. Frames are always allocated in a forward linear fashion. There is no tracking of allocated frames. If your frame gets overwritten, it’s because you held on to it too long. Users of a frame buffer should ensure that the buffer is sufficiently sized to allow external holders of frames to do what they must with them before they are reallocated e.g., frames come with an expiration of one ring epoch. This is essentially a big linear ring of frames. The focus is on fast allocation. Allocations against this frame are expected at rates in the millions to tens of millions per second.

FrameBufferIterator

An iterator over a set of frames. In the event that a series of frames rolls over from the end of the buffer to the beginning this iterator handles that seamlessly.

Ring

A Ring arbitrates access to a frame buffer. Rings are directional, there are producers that allocate and modify and produce elements on one side of a ring and consumers that read and consume frames on the other side of a ring. A ring contains ring elements which are essentially a wrapper around a FrameBuffer specific address.

RingConsumer

A RingConsumer provides read access to a Ring.

RingElement

A wring element is a wrapper around an address to a Frame within a FrameBuffer.

RingIterator

An iterator over a set of RingElements. In the event that a series of elements rolls over from the end of the ring to the beginning, this iterator handles that seamlessly.

RingProducer

A RingProducer provides write access to a Ring.

Enums

Error

Functions

ring

The ring function creates a Ring object and returns a RingProducer and RingConsumer instance for interacting with the Ring.